Cleaning up proptypes
I'm still having trouble with templates and using the Gatsby Head API. I had to defer the default export until after the Head function, otherwise it wouldn't trigger.

I'm getting a Head is not defined error when trying to declare proptypes for the Head function. I think it has something to do with React Fast Refresh.