Update events.js
The "What's happening" panel was not displaying upcoming events. The old solution to compare dates wasn't working. Since both eventDate and now are date objects, we can simplify with a > check to compare the dates and times. The previous code was checking for events today and after, including those that have already concluded from today.