Changed feature list ordering
I was flatmapping the list of features by looking at each feature in each study space. This will pull the order of features in which they appeared. Since bookable appeared last, it was last in the list of features. However, in the CMS we have bookable first. Instead, I recreated the order from the CMS and flatmap against that.